Adiciona estilizaçao css e organiza os componentes
