+++ src/Pages/Home/home.css
@@ -1,0 +1,118 @@
+header {
+  display: flex;
+  padding: 20px;
+}
+
+.header-categories-nav {
+  display: flex;
+  text-align: center;
+}
+
+.header-categories-nav select {
+  background-color: white;
+  border: 1px solid black;
+  border-radius: 20px;
+  padding: 5px;
+}
+
+.header-categories-nav select:focus {
+  box-shadow: 0;
+  outline: 0;
+}
+
+.nav-search-header {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  flex-grow: 2;
+}
+
+.nav-search-input {
+  align-items: center;
+  border: 1px solid black;
+  border-radius: 20px;
+  display: flex;
+  justify-content: space-between;
+  padding: 0 10px;
+  width: 80%;
+}
+
+.nav-search-input input {
+  border: none;
+  padding: 5px;
+  width: 100%;
+}
+
+.nav-search-input input:focus {
+  box-shadow: 0;
+  outline: 0;
+}
+
+.nav-search-input button {
+  background-color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.header-shopping-cart-button {
+  display: flex;
+}
+
+main {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.product-card {
+  border: 1px solid black;
+  border-radius: 20px;
+  height: 510px;
+  margin: 5px;
+  padding: 10px;
+  width: 290px;
+}
+
+.product-card-link {
+  align-items: center;
+  color: black;
+  display: flex;
+  flex-direction: column;
+  height: 480px;
+  justify-content: space-between;
+  text-decoration: none;
+}
+
+.product-img img {
+  height: 290px;
+  width: 290px;
+}
+
+.product-info {
+  flex-grow: 2;
+}
+
+.product-info p {
+  font-size: 24px;
+  font-weight: 400;
+}
+
+.product-info h1 {
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 1.3;
+}
+
+.product-add-to-cart {
+  background-color: rgb(157, 211, 157);
+  border: 1px solid rgb(157, 211, 157);
+  border-radius: 20px;
+  cursor: pointer;
+  padding: 5px;
+  width: 100%;
+}
+
+.product-add-to-cart:hover {
+  background-color: rgb(72, 179, 72);
+  border: 1px solid rgb(72, 179, 72);
+}
